fixing problems in python package
