+++ cpp/s_gd2/__init__.py
@@ -1,0 +1,1 @@
+from .s_gd2 import *
