+++ cpp/s_gd2/swig/layout_wrap.cxx
@@ -1,6 +1,6 @@
 /* ----------------------------------------------------------------------------
  * This file was automatically generated by SWIG (http://www.swig.org).
- * Version 3.0.8
+ * Version 3.0.12
  *
  * This file is not intended to be easily readable and contains a number of
  * coding conventions designed to improve portability and efficiency. Do not make
@@ -106,9 +106,11 @@ template <typename T> T SwigValueInit() {
 #endif
 
 /* exporting methods */
-#if (__GNUC__ >= 4) || (__GNUC__ == 3 && __GNUC_MINOR__ >= 4)
-#  ifndef GCC_HASCLASSVISIBILITY
-#    define GCC_HASCLASSVISIBILITY
+#if defined(__GNUC__)
+#  if (__GNUC__ >= 4) || (__GNUC__ == 3 && __GNUC_MINOR__ >= 4)
+#    ifndef GCC_HASCLASSVISIBILITY
+#      define GCC_HASCLASSVISIBILITY
+#    endif
 #  endif
 #endif
 
@@ -668,16 +670,16 @@ SWIG_UnpackData(const char *c, void *ptr, size_t sz) {
     char d = *(c++);
     unsigned char uu;
     if ((d >= '0') && (d <= '9'))
-      uu = ((d - '0') << 4);
+      uu = (unsigned char)((d - '0') << 4);
     else if ((d >= 'a') && (d <= 'f'))
-      uu = ((d - ('a'-10)) << 4);
+      uu = (unsigned char)((d - ('a'-10)) << 4);
     else
       return (char *) 0;
     d = *(c++);
     if ((d >= '0') && (d <= '9'))
-      uu |= (d - '0');
+      uu |= (unsigned char)(d - '0');
     else if ((d >= 'a') && (d <= 'f'))
-      uu |= (d - ('a'-10));
+      uu |= (unsigned char)(d - ('a'-10));
     else
       return (char *) 0;
     *u = uu;
@@ -860,10 +862,6 @@ PyString_FromFormat(const char *fmt, ...) {
 }
 #endif
 
-/* Add PyObject_Del for old Pythons */
-#if PY_VERSION_HEX < 0x01060000
-# define PyObject_Del(op) PyMem_DEL((op))
-#endif
 #ifndef PyObject_DEL
 # define PyObject_DEL PyObject_Del
 #endif
@@ -978,6 +976,7 @@ typedef destructor freefunc;
 #if PY_VERSION_HEX < 0x03020000
 #define PyDescr_TYPE(x) (((PyDescrObject *)(x))->d_type)
 #define PyDescr_NAME(x) (((PyDescrObject *)(x))->d_name)
+#define Py_hash_t long
 #endif
 
 /* -----------------------------------------------------------------------------
@@ -1926,7 +1925,6 @@ SwigPyObject_TypeOnce(void) {
   static int type_init = 0;
   if (!type_init) {
     const PyTypeObject tmp = {
-      /* PyObject header changed in Python 3 */
 #if PY_VERSION_HEX >= 0x03000000
       PyVarObject_HEAD_INIT(NULL, 0)
 #else
@@ -1937,7 +1935,7 @@ SwigPyObject_TypeOnce(void) {
       sizeof(SwigPyObject),                 /* tp_basicsize */
       0,                                    /* tp_itemsize */
       (destructor)SwigPyObject_dealloc,     /* tp_dealloc */
-      0,				    /* tp_print */
+      0,                                    /* tp_print */
 #if PY_VERSION_HEX < 0x02020000
       (getattrfunc)SwigPyObject_getattr,    /* tp_getattr */
 #else
@@ -1945,7 +1943,7 @@ SwigPyObject_TypeOnce(void) {
 #endif
       (setattrfunc)0,                       /* tp_setattr */
 #if PY_VERSION_HEX >= 0x03000000
-    0, /* tp_reserved in 3.0.1, tp_compare in 3.0.0 but not used */
+      0, /* tp_reserved in 3.0.1, tp_compare in 3.0.0 but not used */
 #else
       (cmpfunc)SwigPyObject_compare,        /* tp_compare */
 #endif
@@ -1955,7 +1953,7 @@ SwigPyObject_TypeOnce(void) {
       0,                                    /* tp_as_mapping */
       (hashfunc)0,                          /* tp_hash */
       (ternaryfunc)0,                       /* tp_call */
-      0,				    /* tp_str */
+      0,                                    /* tp_str */
       PyObject_GenericGetAttr,              /* tp_getattro */
       0,                                    /* tp_setattro */
       0,                                    /* tp_as_buffer */
@@ -2118,7 +2116,6 @@ SwigPyPacked_TypeOnce(void) {
   static int type_init = 0;
   if (!type_init) {
     const PyTypeObject tmp = {
-      /* PyObject header changed in Python 3 */
 #if PY_VERSION_HEX>=0x03000000
       PyVarObject_HEAD_INIT(NULL, 0)
 #else
@@ -3032,7 +3029,7 @@ static swig_module_info swig_module = {swig_types, 1, 0, 0, 0, 0};
 #endif
 #define SWIG_name    "_layout"
 
-#define SWIGVERSION 0x030008 
+#define SWIGVERSION 0x030012 
 #define SWIG_VERSION SWIGVERSION
 
 
@@ -3114,7 +3111,7 @@ namespace swig {
 
 
     #define SWIG_FILE_WITH_INIT
-    #include "layout.hpp"
+    #include "../layout.hpp"
 
 
 #ifndef SWIG_FILE_WITH_INIT
@@ -3149,65 +3146,65 @@ namespace swig {
     void np_layout_unweighted(double* X, int n, int kd,
                               int* I, int len_I,
                               int* J, int len_J,
-                              int t_max, double eps) {
+                              int t_max, double eps, int rseed) {
 
         dimension_check(kd);
         unweighted_edge_check(len_I, len_J);
-        layout_unweighted(n, X, len_I, I, J, t_max, eps);
+        layout_unweighted(n, X, len_I, I, J, t_max, eps, rseed);
     }
     void np_layout_weighted(double* X, int n, int kd,
                             int* I, int len_I,
                             int* J, int len_J,
                             double* V, int len_V,
-                            int t_max, double eps) {
+                            int t_max, double eps, int rseed) {
 
         dimension_check(kd);
         weighted_edge_check(len_I, len_J, len_V);
-        layout_weighted(n, X, len_I, I, J, V, t_max, eps);
+        layout_weighted(n, X, len_I, I, J, V, t_max, eps, rseed);
     }
     void np_layout_unweighted_convergent(double* X, int n, int kd,
                                          int* I, int len_I,
                                          int* J, int len_J,
-                                         int t_max, double eps, double delta, int t_maxmax) {
+                                         int t_max, double eps, double delta, int t_maxmax, int rseed) {
 
         dimension_check(kd);
         unweighted_edge_check(len_I, len_J);
-        layout_unweighted_convergent(n, X, len_I, I, J, t_max, eps, delta, t_maxmax);
+        layout_unweighted_convergent(n, X, len_I, I, J, t_max, eps, delta, t_maxmax, rseed);
     }
     void np_layout_weighted_convergent(double* X, int n, int kd,
                                        int* I, int len_I,
                                        int* J, int len_J,
                                        double* V, int len_V,
-                                       int t_max, double eps, double delta, int t_maxmax) {
+                                       int t_max, double eps, double delta, int t_maxmax, int rseed) {
 
         dimension_check(kd);
         weighted_edge_check(len_I, len_J, len_V);
-        layout_weighted_convergent(n, X, len_I, I, J, V, t_max, eps, delta, t_maxmax);
+        layout_weighted_convergent(n, X, len_I, I, J, V, t_max, eps, delta, t_maxmax, rseed);
     }
     void np_layout_sparse_unweighted(double* X, int n, int kd,
                                      int* I, int len_I,
                                      int* J, int len_J,
-                                     int p, int t_max, double eps) {
+                                     int p, int t_max, double eps, int rseed) {
 
         dimension_check(kd);
         unweighted_edge_check(len_I, len_J);
-        layout_sparse_unweighted(n, X, len_I, I, J, p, t_max, eps);
+        layout_sparse_unweighted(n, X, len_I, I, J, p, t_max, eps, rseed);
     }
     void np_layout_sparse_weighted(double* X, int n, int kd,
                                      int* I, int len_I,
                                      int* J, int len_J,
                                      double* V, int len_V,
-                                     int p, int t_max, double eps) {
+                                     int p, int t_max, double eps, int rseed) {
 
         dimension_check(kd);
         weighted_edge_check(len_I, len_J, len_V);
-        layout_sparse_weighted(n, X, len_I, I, J, V, p, t_max, eps);
+        layout_sparse_weighted(n, X, len_I, I, J, V, p, t_max, eps, rseed);
     }
 
     void np_mds_direct(double* X, int n, int kd,
                        double* d, int len_d,
                        double* w, int len_w,
-                       double* eta, int len_eta) {
+                       double* eta, int len_eta, int rseed) {
 
         dimension_check(kd);
         int nC2 = (n*(n-1))/2;
@@ -3215,7 +3212,7 @@ namespace swig {
             PyErr_Format(PyExc_ValueError, "d or w not right length for condensed distance matrix");
             return;
         }
-        mds_direct(n, X, d, w, len_eta, eta);
+        mds_direct(n, X, d, w, len_eta, eta, rseed);
     }
 
 
@@ -3238,7 +3235,7 @@ SWIG_AsVal_double (PyObject *obj, double *val)
     return SWIG_OK;
 #if PY_VERSION_HEX < 0x03000000
   } else if (PyInt_Check(obj)) {
-    if (val) *val = PyInt_AsLong(obj);
+    if (val) *val = (double) PyInt_AsLong(obj);
     return SWIG_OK;
 #endif
   } else if (PyLong_Check(obj)) {
@@ -3951,6 +3948,7 @@ SWIGINTERN PyObject *_wrap_layout_unweighted(PyObject *SWIGUNUSEDPARM(self), PyO
   int arg7 ;
   int arg8 ;
   double arg9 ;
+  int arg10 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -3960,13 +3958,16 @@ SWIGINTERN PyObject *_wrap_layout_unweighted(PyObject *SWIGUNUSEDPARM(self), PyO
   int ecode8 = 0 ;
   double val9 ;
   int ecode9 = 0 ;
+  int val10 ;
+  int ecode10 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
   PyObject * obj3 = 0 ;
   PyObject * obj4 = 0 ;
+  PyObject * obj5 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOO:layout_unweighted",&obj0,&obj1,&obj2,&obj3,&obj4)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOO:layout_unweighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4009,8 +4010,13 @@ SWIGINTERN PyObject *_wrap_layout_unweighted(PyObject *SWIGUNUSEDPARM(self), PyO
     SWIG_exception_fail(SWIG_ArgError(ecode9), "in method '" "layout_unweighted" "', argument " "9"" of type '" "double""'");
   } 
   arg9 = static_cast< double >(val9);
+  ecode10 = SWIG_AsVal_int(obj5, &val10);
+  if (!SWIG_IsOK(ecode10)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode10), "in method '" "layout_unweighted" "', argument " "10"" of type '" "int""'");
+  } 
+  arg10 = static_cast< int >(val10);
   {
-    np_layout_unweighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9);
+    np_layout_unweighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4057,6 +4063,7 @@ SWIGINTERN PyObject *_wrap_layout_weighted(PyObject *SWIGUNUSEDPARM(self), PyObj
   int arg9 ;
   int arg10 ;
   double arg11 ;
+  int arg12 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4068,14 +4075,17 @@ SWIGINTERN PyObject *_wrap_layout_weighted(PyObject *SWIGUNUSEDPARM(self), PyObj
   int ecode10 = 0 ;
   double val11 ;
   int ecode11 = 0 ;
+  int val12 ;
+  int ecode12 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
   PyObject * obj3 = 0 ;
   PyObject * obj4 = 0 ;
   PyObject * obj5 = 0 ;
+  PyObject * obj6 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOOO:layout_weighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOOO:layout_weighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4130,8 +4140,13 @@ SWIGINTERN PyObject *_wrap_layout_weighted(PyObject *SWIGUNUSEDPARM(self), PyObj
     SWIG_exception_fail(SWIG_ArgError(ecode11), "in method '" "layout_weighted" "', argument " "11"" of type '" "double""'");
   } 
   arg11 = static_cast< double >(val11);
+  ecode12 = SWIG_AsVal_int(obj6, &val12);
+  if (!SWIG_IsOK(ecode12)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode12), "in method '" "layout_weighted" "', argument " "12"" of type '" "int""'");
+  } 
+  arg12 = static_cast< int >(val12);
   {
-    np_layout_weighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11);
+    np_layout_weighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4190,6 +4205,7 @@ SWIGINTERN PyObject *_wrap_layout_unweighted_convergent(PyObject *SWIGUNUSEDPARM
   double arg9 ;
   double arg10 ;
   int arg11 ;
+  int arg12 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4203,6 +4219,8 @@ SWIGINTERN PyObject *_wrap_layout_unweighted_convergent(PyObject *SWIGUNUSEDPARM
   int ecode10 = 0 ;
   int val11 ;
   int ecode11 = 0 ;
+  int val12 ;
+  int ecode12 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
@@ -4210,8 +4228,9 @@ SWIGINTERN PyObject *_wrap_layout_unweighted_convergent(PyObject *SWIGUNUSEDPARM
   PyObject * obj4 = 0 ;
   PyObject * obj5 = 0 ;
   PyObject * obj6 = 0 ;
+  PyObject * obj7 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOOOO:layout_unweighted_convergent",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOOOO:layout_unweighted_convergent",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6,&obj7)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4264,8 +4283,13 @@ SWIGINTERN PyObject *_wrap_layout_unweighted_convergent(PyObject *SWIGUNUSEDPARM
     SWIG_exception_fail(SWIG_ArgError(ecode11), "in method '" "layout_unweighted_convergent" "', argument " "11"" of type '" "int""'");
   } 
   arg11 = static_cast< int >(val11);
+  ecode12 = SWIG_AsVal_int(obj7, &val12);
+  if (!SWIG_IsOK(ecode12)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode12), "in method '" "layout_unweighted_convergent" "', argument " "12"" of type '" "int""'");
+  } 
+  arg12 = static_cast< int >(val12);
   {
-    np_layout_unweighted_convergent(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11);
+    np_layout_unweighted_convergent(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4314,6 +4338,7 @@ SWIGINTERN PyObject *_wrap_layout_weighted_convergent(PyObject *SWIGUNUSEDPARM(s
   double arg11 ;
   double arg12 ;
   int arg13 ;
+  int arg14 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4329,6 +4354,8 @@ SWIGINTERN PyObject *_wrap_layout_weighted_convergent(PyObject *SWIGUNUSEDPARM(s
   int ecode12 = 0 ;
   int val13 ;
   int ecode13 = 0 ;
+  int val14 ;
+  int ecode14 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
@@ -4337,8 +4364,9 @@ SWIGINTERN PyObject *_wrap_layout_weighted_convergent(PyObject *SWIGUNUSEDPARM(s
   PyObject * obj5 = 0 ;
   PyObject * obj6 = 0 ;
   PyObject * obj7 = 0 ;
+  PyObject * obj8 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOOOOO:layout_weighted_convergent",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6,&obj7)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOOOOO:layout_weighted_convergent",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6,&obj7,&obj8)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4403,8 +4431,13 @@ SWIGINTERN PyObject *_wrap_layout_weighted_convergent(PyObject *SWIGUNUSEDPARM(s
     SWIG_exception_fail(SWIG_ArgError(ecode13), "in method '" "layout_weighted_convergent" "', argument " "13"" of type '" "int""'");
   } 
   arg13 = static_cast< int >(val13);
+  ecode14 = SWIG_AsVal_int(obj8, &val14);
+  if (!SWIG_IsOK(ecode14)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode14), "in method '" "layout_weighted_convergent" "', argument " "14"" of type '" "int""'");
+  } 
+  arg14 = static_cast< int >(val14);
   {
-    np_layout_weighted_convergent(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12,arg13);
+    np_layout_weighted_convergent(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12,arg13,arg14);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4462,6 +4495,7 @@ SWIGINTERN PyObject *_wrap_layout_sparse_unweighted(PyObject *SWIGUNUSEDPARM(sel
   int arg8 ;
   int arg9 ;
   double arg10 ;
+  int arg11 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4473,14 +4507,17 @@ SWIGINTERN PyObject *_wrap_layout_sparse_unweighted(PyObject *SWIGUNUSEDPARM(sel
   int ecode9 = 0 ;
   double val10 ;
   int ecode10 = 0 ;
+  int val11 ;
+  int ecode11 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
   PyObject * obj3 = 0 ;
   PyObject * obj4 = 0 ;
   PyObject * obj5 = 0 ;
+  PyObject * obj6 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOOO:layout_sparse_unweighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOOO:layout_sparse_unweighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4528,8 +4565,13 @@ SWIGINTERN PyObject *_wrap_layout_sparse_unweighted(PyObject *SWIGUNUSEDPARM(sel
     SWIG_exception_fail(SWIG_ArgError(ecode10), "in method '" "layout_sparse_unweighted" "', argument " "10"" of type '" "double""'");
   } 
   arg10 = static_cast< double >(val10);
+  ecode11 = SWIG_AsVal_int(obj6, &val11);
+  if (!SWIG_IsOK(ecode11)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode11), "in method '" "layout_sparse_unweighted" "', argument " "11"" of type '" "int""'");
+  } 
+  arg11 = static_cast< int >(val11);
   {
-    np_layout_sparse_unweighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10);
+    np_layout_sparse_unweighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4577,6 +4619,7 @@ SWIGINTERN PyObject *_wrap_layout_sparse_weighted(PyObject *SWIGUNUSEDPARM(self)
   int arg10 ;
   int arg11 ;
   double arg12 ;
+  int arg13 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4590,6 +4633,8 @@ SWIGINTERN PyObject *_wrap_layout_sparse_weighted(PyObject *SWIGUNUSEDPARM(self)
   int ecode11 = 0 ;
   double val12 ;
   int ecode12 = 0 ;
+  int val13 ;
+  int ecode13 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
@@ -4597,8 +4642,9 @@ SWIGINTERN PyObject *_wrap_layout_sparse_weighted(PyObject *SWIGUNUSEDPARM(self)
   PyObject * obj4 = 0 ;
   PyObject * obj5 = 0 ;
   PyObject * obj6 = 0 ;
+  PyObject * obj7 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOOOOO:layout_sparse_weighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOOOOO:layout_sparse_weighted",&obj0,&obj1,&obj2,&obj3,&obj4,&obj5,&obj6,&obj7)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4658,8 +4704,13 @@ SWIGINTERN PyObject *_wrap_layout_sparse_weighted(PyObject *SWIGUNUSEDPARM(self)
     SWIG_exception_fail(SWIG_ArgError(ecode12), "in method '" "layout_sparse_weighted" "', argument " "12"" of type '" "double""'");
   } 
   arg12 = static_cast< double >(val12);
+  ecode13 = SWIG_AsVal_int(obj7, &val13);
+  if (!SWIG_IsOK(ecode13)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode13), "in method '" "layout_sparse_weighted" "', argument " "13"" of type '" "int""'");
+  } 
+  arg13 = static_cast< int >(val13);
   {
-    np_layout_sparse_weighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12);
+    np_layout_sparse_weighted(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10,arg11,arg12,arg13);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -4716,6 +4767,7 @@ SWIGINTERN PyObject *_wrap_mds_direct(PyObject *SWIGUNUSEDPARM(self), PyObject *
   int arg7 ;
   double *arg8 = (double *) 0 ;
   int arg9 ;
+  int arg10 ;
   PyArrayObject *array1 = NULL ;
   PyArrayObject *array4 = NULL ;
   int is_new_object4 = 0 ;
@@ -4723,12 +4775,15 @@ SWIGINTERN PyObject *_wrap_mds_direct(PyObject *SWIGUNUSEDPARM(self), PyObject *
   int is_new_object6 = 0 ;
   PyArrayObject *array8 = NULL ;
   int is_new_object8 = 0 ;
+  int val10 ;
+  int ecode10 = 0 ;
   PyObject * obj0 = 0 ;
   PyObject * obj1 = 0 ;
   PyObject * obj2 = 0 ;
   PyObject * obj3 = 0 ;
+  PyObject * obj4 = 0 ;
   
-  if (!PyArg_ParseTuple(args,(char *)"OOOO:mds_direct",&obj0,&obj1,&obj2,&obj3)) SWIG_fail;
+  if (!PyArg_ParseTuple(args,(char *)"OOOOO:mds_direct",&obj0,&obj1,&obj2,&obj3,&obj4)) SWIG_fail;
   {
     array1 = obj_to_array_no_conversion(obj0, NPY_DOUBLE);
     if (!array1 || !require_dimensions(array1,2) || !require_contiguous(array1)
@@ -4773,8 +4828,13 @@ SWIGINTERN PyObject *_wrap_mds_direct(PyObject *SWIGUNUSEDPARM(self), PyObject *
     arg8 = (double*) array_data(array8);
     arg9 = (int) array_size(array8,0);
   }
+  ecode10 = SWIG_AsVal_int(obj4, &val10);
+  if (!SWIG_IsOK(ecode10)) {
+    SWIG_exception_fail(SWIG_ArgError(ecode10), "in method '" "mds_direct" "', argument " "10"" of type '" "int""'");
+  } 
+  arg10 = static_cast< int >(val10);
   {
-    np_mds_direct(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9);
+    np_mds_direct(arg1,arg2,arg3,arg4,arg5,arg6,arg7,arg8,arg9,arg10);
     if (PyErr_Occurred()) SWIG_fail;
   }
   resultobj = SWIG_Py_Void();
@@ -5227,7 +5287,6 @@ extern "C" {
     static int type_init = 0;
     if (!type_init) {
       const PyTypeObject tmp = {
-        /* PyObject header changed in Python 3 */
 #if PY_VERSION_HEX >= 0x03000000
         PyVarObject_HEAD_INIT(NULL, 0)
 #else
@@ -5459,13 +5518,13 @@ SWIG_init(void) {
   static PyGetSetDef thisown_getset_def = {
     (char *)"thisown", SwigPyBuiltin_GetterClosure, SwigPyBuiltin_SetterClosure, NULL, &thisown_getset_closure
   };
-  PyObject *metatype_args;
   PyTypeObject *builtin_pytype;
   int builtin_base_count;
   swig_type_info *builtin_basetype;
   PyObject *tuple;
   PyGetSetDescrObject *static_getset;
   PyTypeObject *metatype;
+  PyTypeObject *swigpyobject;
   SwigPyClientData *cd;
   PyObject *public_interface, *public_symbol;
   PyObject *this_descr;
@@ -5480,14 +5539,9 @@ SWIG_init(void) {
   (void)static_getset;
   (void)self;
   
-  /* metatype is used to implement static member variables. */
-  metatype_args = Py_BuildValue("(s(O){})", "SwigPyObjectType", &PyType_Type);
-  assert(metatype_args);
-  metatype = (PyTypeObject *) PyType_Type.tp_call((PyObject *) &PyType_Type, metatype_args, NULL);
+  /* Metaclass is used to implement static member variables */
+  metatype = SwigPyObjectType();
   assert(metatype);
-  Py_DECREF(metatype_args);
-  metatype->tp_setattro = (setattrofunc) &SwigPyObjectType_setattro;
-  assert(PyType_Ready(metatype) >= 0);
 #endif
   
   /* Fix SwigMethods to carry the callback ptrs when needed */
@@ -5505,13 +5559,15 @@ SWIG_init(void) {
   SWIG_InitializeModule(0);
   
 #ifdef SWIGPYTHON_BUILTIN
+  swigpyobject = SwigPyObject_TypeOnce();
+  
   SwigPyObject_stype = SWIG_MangledTypeQuery("_p_SwigPyObject");
   assert(SwigPyObject_stype);
   cd = (SwigPyClientData*) SwigPyObject_stype->clientdata;
   if (!cd) {
     SwigPyObject_stype->clientdata = &SwigPyObject_clientdata;
-    SwigPyObject_clientdata.pytype = SwigPyObject_TypeOnce();
-  } else if (SwigPyObject_TypeOnce()->tp_basicsize != cd->pytype->tp_basicsize) {
+    SwigPyObject_clientdata.pytype = swigpyobject;
+  } else if (swigpyobject->tp_basicsize != cd->pytype->tp_basicsize) {
     PyErr_SetString(PyExc_RuntimeError, "Import error: attempted to load two incompatible swig-generated modules.");
 # if PY_VERSION_HEX >= 0x03000000
     return NULL;
